feat: add dedicate layer for scenes and also the first book example
A new layer was add to handle scenes, thoses scenes aims to contain the
code for all the examples in the book.

I just move the GL logic inside core/renderer/System, to handle glad and
webgl/es, for now with pre-processor conditional. Maybe in the future I
develo a better layer for that, since it is not the main objective of
this book/repository.